added  payment component
