+++ src/App.tsx
@@ -153,6 +153,7 @@ import { DrawerUserLookup } from "./components/Drawer/DrawerUserLookup";
 import { UserLookup } from "./components/UserLookup.tsx/UserLookup";
 import { RegisterName } from "./components/RegisterName";
 import { BuyQortInformation } from "./components/BuyQortInformation";
+import { QortPayment } from "./components/QortPayment";
 
 type extStates =
   | "not-authenticated"
@@ -351,6 +352,8 @@ function App() {
   const [authenticatePassword, setAuthenticatePassword] = useState<string>("");
   const [sendqortState, setSendqortState] = useState<any>(null);
   const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [isLoadingSendCoin, setIsLoadingSendCoin] = useState<boolean>(false);
+
   const [
     walletToBeDownloadedPasswordConfirm,
     setWalletToBeDownloadedPasswordConfirm,
@@ -775,52 +778,7 @@ function App() {
         setLtcBalanceLoading(false);
       });
   };
-  const sendCoinFunc = async() => {
-   try {
-    setSendPaymentError("");
-    setSendPaymentSuccess("");
-    if (!paymentTo) {
-      setSendPaymentError("Please enter a recipient");
-      return;
-    }
-    if (!paymentAmount) {
-      setSendPaymentError("Please enter an amount greater than 0");
-      return;
-    }
-    if (!paymentPassword) {
-      setSendPaymentError("Please enter your wallet password");
-      return;
-    }
-    const fee = await getFee('PAYMENT')
-
-    await show({
-      message: `Would you like to transfer ${Number(paymentAmount)} QORT?` ,
-      paymentFee: fee.fee + ' QORT'
-    })
-    setIsLoading(true);
-    window
-      .sendMessage("sendCoin", {
-        amount: Number(paymentAmount),
-        receiver: paymentTo.trim(),
-        password: paymentPassword,
-      })
-      .then((response) => {
-        if (response?.error) {
-          setSendPaymentError(response.error);
-        } else {
-          setIsOpenSendQort(false);
-          setIsOpenSendQortSuccess(true);
-        }
-        setIsLoading(false);
-      })
-      .catch((error) => {
-        console.error("Failed to send coin:", error);
-        setIsLoading(false);
-      });
-   } catch (error) {
-    // error
-   }
-  };
+ 
 
   const clearAllStates = () => {
     setRequestConnection(null);
@@ -2119,94 +2077,12 @@ function App() {
             />
           </Box>
           <Spacer height="35px" />
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "column",
-              alignItems: "flex-start",
-            }}
-          >
-            <TextP
-              sx={{
-                textAlign: "start",
-                lineHeight: "24px",
-                fontSize: "20px",
-                fontWeight: 600,
-              }}
-            >
-              Transfer QORT
-            </TextP>
-            <Spacer height="35px" />
-            <TextP
-              sx={{
-                textAlign: "start",
-                lineHeight: "16px",
-                fontSize: "20px",
-                fontWeight: 600,
-                color: "rgba(255, 255, 255, 0.5)",
-              }}
-            >
-              Balance:
-            </TextP>
-            <TextP
-              sx={{
-                textAlign: "start",
-                lineHeight: "24px",
-                fontSize: "20px",
-                fontWeight: 700,
-              }}
-            >
-              {balance?.toFixed(2)} QORT
-            </TextP>
-          </Box>
-          <Spacer height="35px" />
-
-          <Box>
-            <CustomLabel htmlFor="standard-adornment-name">To</CustomLabel>
-            <Spacer height="5px" />
-            <CustomInput
-              id="standard-adornment-name"
-              value={paymentTo}
-              onChange={(e) => setPaymentTo(e.target.value)}
-              autoComplete="off"
+            <QortPayment balance={balance} show={show} onSuccess={()=> {
+               setIsOpenSendQort(false);
+               setIsOpenSendQortSuccess(true);
+            }} 
+             defaultPaymentTo={paymentTo}
             />
-            <Spacer height="6px" />
-            <CustomLabel htmlFor="standard-adornment-amount">
-              Amount
-            </CustomLabel>
-            <Spacer height="5px" />
-             <BoundedNumericTextField
-              value={paymentAmount}
-              minValue={0}
-               maxValue={+balance}
-                allowDecimals={true}
-                initialValue={'0'}
-                allowNegatives={false}
-                afterChange={(e: string) => setPaymentAmount(+e)}
-            />
-            <Spacer height="6px" />
-            <CustomLabel htmlFor="standard-adornment-password">
-              Confirm Wallet Password
-            </CustomLabel>
-            <Spacer height="5px" />
-            <PasswordField
-              id="standard-adornment-password"
-              value={paymentPassword}
-              onChange={(e) => setPaymentPassword(e.target.value)}
-              autoComplete="off"
-            />
-          </Box>
-          <Spacer height="10px" />
-          <ErrorText>{sendPaymentError}</ErrorText>
-          {/* <Typography>{sendPaymentSuccess}</Typography> */}
-          <Spacer height="25px" />
-          <CustomButton
-            onClick={() => {
-              sendCoinFunc();
-            }}
-          >
-            Send
-          </CustomButton>
         </Box>
       )}
